% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimaModelFit.R
\name{arimaModelFit}
\alias{arimaModelFit}
\title{Measures of model fit for to assess ARIMA models}
\usage{
arimaModelFit(arimaEstimation, decim = 4)
}
\arguments{
\item{arimaEstimation}{An object of class arima.}

\item{decim}{Specifies the number of decimals to export results. The default is 4.}
}
\value{
A dataframe with Log likelihood, Akaike information criterion, Bayesian information criterion, Invertible AR roots, Invertible MA roots
}
\description{
Measures of model fit for to assess ARIMA models
}
\examples{
library(quantmod)
library(urca)
getSymbols("AAPL", src="yahoo")
type=unitRootDF_ABsequential(AAPL$AAPL.Close)
aapl_stationary=stationarize(ts(AAPL$AAPL.Close), type=type)
arimaEstimation=arima(aapl_stationary, order = c(1,0,0))
arimaModelFit(arimaEstimation, decim=2)
}
