% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastOriginal.r
\name{forecastOriginal}
\alias{forecastOriginal}
\title{Takes an ARIMA estimation on stationary and seasonally adjusted series and returns the forecast on the original, non-stationary, not seasonally adjusted series}
\usage{
forecastOriginal(
  arimaEstimation,
  type,
  seasonalCoeff,
  seasonalityType,
  nQuarters = 4,
  year,
  quarter,
  saSeries
)
}
\arguments{
\item{arimaEstimation}{An object of class arima.}

\item{type}{Specifies the type of the original time series.
The options are: "Stationary around a linear trend", "Random walk with drift", "Unit root and linear trend"}

\item{seasonalCoeff}{A vector of seasonal coefficients.}

\item{seasonalityType}{"additive", "multiplicative", or "none".}

\item{nQuarters}{Number of quarters to forecast. Default is 4.}

\item{year}{The year when the forecast should start.}

\item{quarter}{The quarter when the forecast should start.}

\item{saSeries}{A vector with the seasonally adjusted series.}
}
\value{
The forecasted values for the original non-stationary, not seasonally adjusted series.
}
\description{
Takes an ARIMA estimation on stationary and seasonally adjusted series and returns the forecast on the original, non-stationary, not seasonally adjusted series
}
\examples{
library(urca)
library (forecast)
library(wooldridge)
data('intqrt')
plot(intqrt$r3, type="l")
Tbill3M=ts(as.numeric(intqrt$r3), start=c(1960,1), frequency=4)
Tbill3M_dec=decompose(Tbill3M, type ="multiplicative")
seasonalCoeff=Tbill3M_dec$figure
Tbill3M_sa=ts(Tbill3M/Tbill3M_dec$seasonal, start=c(1960,1), frequency=4)
type=unitRootDF_ABsequential(Tbill3M_sa)
Tbill3M_stationary=stationarize(ts(Tbill3M_sa, start=c(1960,1), frequency=4), type=type)
arimaEstimation=arima(Tbill3M_stationary, order = c(1,0,0))
forecastOriginal(arimaEstimation, type, seasonalCoeff, seasonalityType="multiplicative", nQuarters=8, year=1991, quarter=1, saSeries=Tbill3M_sa)
}
