% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarize.R
\name{stationarize}
\alias{stationarize}
\title{Takes a time series of type "Stationary around a linear trend", "Random walk with drift", or "Unit root and linear trend" and returns a time series that is "Stationary, no trend"}
\usage{
stationarize(x, type, pvalue = 0.05, ic = "BIC")
}
\arguments{
\item{x}{An object of class ts.}

\item{type}{Specifieces the type of the time series.
The options are: "Stationary around a linear trend", "Random walk with drift", "Unit root and linear trend"}

\item{pvalue}{Specifies the threshold for rejecting the null hypothesis after first differencing. The default is 0.05.}

\item{ic}{Lag selection based on "AIC" or "BIC". The default is "BIC". Used for DF testing after first differencing.}
}
\value{
A stationary, no trend series.
}
\description{
Takes a time series of type "Stationary around a linear trend", "Random walk with drift", or "Unit root and linear trend" and returns a time series that is "Stationary, no trend"
}
\examples{
library(quantmod)
library(urca)
getSymbols("AAPL", src="yahoo")
type=unitRootDF_ABsequential(AAPL$AAPL.Close)
aapl_stationary=stationarize(ts(AAPL$AAPL.Close), type=type)
}
