% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitRootDF_ABsequential.R
\name{unitRootDF_ABsequential}
\alias{unitRootDF_ABsequential}
\title{Dickey-Fuller test applied following the Ayat and Burridge (2000) sequential procedure}
\usage{
unitRootDF_ABsequential(x, pvalue = 0.05, ic = "BIC")
}
\arguments{
\item{x}{An object of class ts.}

\item{pvalue}{Specifies the threshold for rejecting the null hypothesis. The default is 0.05.}

\item{ic}{Lag selection based on "AIC" or "BIC". The default is "BIC".}
}
\value{
A text: "Stationary, no trend", "Stationary around a linear trend", "Random walk with drift", "Unit root and linear trend".
}
\description{
Dickey-Fuller test applied following the Ayat and Burridge (2000) sequential procedure
}
\examples{
library(quantmod)
library(urca)
getSymbols("AAPL", src="yahoo")
unitRootDF_ABsequential(AAPL$AAPL.Close)
unitRootDF_ABsequential(AAPL$AAPL.Close, 0.01)
}
\references{
Ayat, L., Burridge, P. (2000), Unit root tests in the presence of uncertainty about the non-stochastic trend, Journal of Econometrics, Vol. 95, Issue 1, pp. 71-96, DOI: 10.1016/S0304-4076(99)00030-5.
https://cran.r-project.org/web/packages/urca/index.html
}
